requirements.txt
flask==2.3.3
transformers==4.38.0
torch==2.2.0
sentence-transformers==2.5.0
faiss-cpu==1.7.4
numpy==1.26.0
scikit-learn==1.4.0
PyPDF2==3.0.1
werkzeug==2.3.7
tqdm==4.66.1 